+++ src/features/seller/components/index.ts
@@ -2,6 +2,7 @@ export { SellerProductCard } from "./SellerProductCard";
 export { SellerOrdersView } from "./SellerOrdersView";
 export { SellerProductsView } from "./SellerProductsView";
 export { SellerDashboardView } from "./SellerDashboardView";
+export { SellerCreateProductView } from "./SellerCreateProductView";
 export { SellerStatCard } from "./SellerStatCard";
 export { SellerQuickActions } from "./SellerQuickActions";
 export { SellerRecentListings } from "./SellerRecentListings";